+++ src/components/ui/UserPanel.tsx
@@ -1,16 +1,17 @@
-import { AvatarProps, Group, GroupProps, Stack, Text } from '@mantine/core';
+import { AvatarProps, Flex, FlexProps, Group, GroupProps, Stack, Text } from '@mantine/core';
 import { useUser } from '../../hooks/useUser';
 import { IUser } from '../../types/models';
 import UserAvatar from './UserAvatar';
 import UserName from './UserName';
 
-interface UserPanelProps extends GroupProps {
+interface UserPanelProps extends FlexProps {
   user?: IUser;
   description?: string;
   withUsername?: boolean;
   avatarSize?: AvatarProps['size'];
   withName?: boolean;
   noAction?: boolean;
+  extra?: React.ReactNode;
 }
 
 export default function UserPanel({
@@ -20,21 +21,22 @@ export default function UserPanel({
   withName = true,
   avatarSize = 'sm',
   noAction = false,
+  extra,
   ...props
 }: UserPanelProps) {
   const { username } = user ?? useUser();
 
   return (
-    <Group gap={10} {...props}>
+    <Flex align='center' gap={10} {...props}>
       <UserAvatar user={user} size={avatarSize} />
       {withName && (
         <Stack gap={0}>
-          <UserName user={user} noAction={noAction} />
+          <UserName user={user} title={extra} noAction={noAction} />
           <Text size='xs' c='dimmed'>
             {description ? description : withUsername ? `@${username}` : null}
           </Text>
         </Stack>
       )}
-    </Group>
+    </Flex>
   );
 }