+++ src/schemas/post.schema.ts
@@ -5,6 +5,7 @@ export const CreatePostSchema = z.object({
     .string()
     .min(1, 'Post content is required')
     .max(500, 'Maximum content length is 500 characters'),
+  mentions: z.array(z.string()).max(6, 'You cannot mention more than 6 users'),
 });
 
 export type TCreatePostSchema = z.infer<typeof CreatePostSchema>;